BAU - integration-tests - Refactor step definitions to separate out common steps
