+++ integration-tests/build.gradle
@@ -10,7 +10,7 @@ repositories {
 }
 
 ext {
-	cucumberVersion = '7.7.0'
+	cucumberVersion = '7.9.0'
 	junitJupiterapi = '5.9.0'
 	jacksonDatabind = '2.9.5'
 	apacheHttpClient = '4.5.13'
@@ -20,6 +20,7 @@ ext {
 dependencies {
 	testImplementation "io.cucumber:cucumber-java:$cucumberVersion"
 	testImplementation "io.cucumber:cucumber-junit:$cucumberVersion"
+	testImplementation "io.cucumber:cucumber-picocontainer:$cucumberVersion"
 	testImplementation "org.apache.httpcomponents:httpclient:$apacheHttpClient"
 	testImplementation "org.apache.commons:commons-lang3:$apacheCommonLang"
 	testImplementation "org.junit.jupiter:junit-jupiter-api:$junitJupiterapi"
